[refactor] Avoid dist folder in sam-highlighter
Without it, we can freely import submodules inside `sam-highlighter`
