+++ sam-highlighter/language.d.ts
@@ -1,0 +1,2 @@
+declare const _default: () => void;
+export default _default;
