--- sam-highlighter/src/index.ts
@@ -1,2 +1,0 @@
-export { default } from './CodeBlock';
-export { default as initialize } from './language';
